requirements.txt
numpy==1.22.3
pandas==1.5.2
matplotlib==3.6.2
seaborn==0.12.2
tqdm==4.64.1
scikit-learn==1.2.1
imbalanced-learn==0.9.1
xgboost==1.7.1
lightgbm==3.3.3
torch==1.13.1
torchvision==0.14.1
pytorch-lightning==1.8.6
timm==0.6.12
hydra-core==1.3.1
einops==0.3.2
pickle5==0.0.11
requests==2.28.1
urllib3==1.26.14
chardet==4.0.0
polars==0.15.15
pyarrow==10.0.1
cycler==0.11.0
kiwisolver==1.4.4
omegaconf==2.3.0
openpack-toolkit==0.7.0
pyparsing==3.0.9
hmmlearn==0.3.0
fastdtw==0.3.4
statsmodels==0.13.5